[cyclic-import] Break cycle between pylint.checkers.util / variables
