+++ pylint/checkers/variables.py
@@ -15,7 +15,6 @@ import sys
 from collections import defaultdict
 from collections.abc import Generator, Iterable, Iterator
 from enum import Enum
-from functools import lru_cache
 from typing import TYPE_CHECKING, Any, NamedTuple
 
 import astroid
@@ -28,6 +27,7 @@ from pylint.checkers.utils import (
     in_type_checking_block,
     is_postponed_evaluation_enabled,
     is_sys_guard,
+    overridden_method,
 )
 from pylint.constants import PY39_PLUS, TYPING_NEVER, TYPING_NORETURN
 from pylint.interfaces import CONTROL_FLOW, HIGH, INFERENCE, INFERENCE_FAILURE
@@ -148,26 +148,6 @@ def _is_from_future_import(stmt: nodes.ImportFrom, name: str) -> bool | None:
     for local_node in module.locals.get(name, []):
         if isinstance(local_node, nodes.ImportFrom) and local_node.modname == FUTURE:
             return True
-    return None
-
-
-@lru_cache(maxsize=1000)
-def overridden_method(
-    klass: nodes.LocalsDictNodeNG, name: str | None
-) -> nodes.FunctionDef | None:
-    """Get overridden method if any."""
-    try:
-        parent = next(klass.local_attr_ancestors(name))
-    except (StopIteration, KeyError):
-        return None
-    try:
-        meth_node = parent[name]
-    except KeyError:
-        # We have found an ancestor defining <name> but it's not in the local
-        # dictionary. This may happen with astroid built from living objects.
-        return None
-    if isinstance(meth_node, nodes.FunctionDef):
-        return meth_node
     return None
